{{cookiecutter.repo_name}}/requirements.txt
django>=1.4.5
django-model-utils>=1.4.0